+++ src/helpers/blobs.ts
@@ -1,0 +1,3 @@
+export function convertBlobToFile(blob: Blob, fileName: string): File {
+    return new File([blob], fileName, { type: "audio/wav" });
+}
