+++ src/pages/Home.tsx
@@ -3,6 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { useGetGames } from "../hooks/queries.ts";
 import HomeCard from "../components/common/cards/HomeCard.tsx";
 import SpinnerLoader from "../components/common/SpinnerLoader.tsx";
+import { useDispatch } from "react-redux";
+import { selectGame } from "../redux/store/gameSlice.ts";
 
 const getCardBgColor = (index: number) => {
   const colors = ["bg-blue-300", "bg-orange-300", "bg-orange-150"];
@@ -11,6 +13,7 @@ const getCardBgColor = (index: number) => {
 
 const Home: React.FC = () => {
   const navigate = useNavigate();
+  const dispatch = useDispatch();
   const { games, isLoading, error } = useGetGames();
 
   if (error) {
@@ -24,7 +27,7 @@ const Home: React.FC = () => {
   return games && games.length !== 0 && !error ? (
     <div className="w-full grid grid-cols-1 md:grid-cols-3 mt-16 gap-10">
       {games.map((game) => (
-        <div key={game.id} className="flex justify-center items-center w-full">
+        <div key={game.id} className="flex justify-center items-center w-full" onClick={() => dispatch(selectGame(game))}>
           <HomeCard
             buttonVariant="secondary"
             onClick={() => navigate(`/actividad/${game.id}/tematicas`)}
